OPS-2361 Add command diff
This implements a command diff that diffs a running stack's template via
the GetTemplate API and the template generated by the stack.

The diff method is implemented in the Stack class in order to have
access to the stack's external_name.

Unfortunately, the Stack class does not have access to the Plan, making
it awkward to re-use the recently-implemented fetch_remote_template
command.