Suppress stack name in output
For consistency with the fetch-remote-template and generate commands, I
have modified the diff command to NOT show the stack name. While this
is leads to what is arguably confusing output in the case of a stack
group with multiple stacks, it seems that upholding consistency in
design is more important.